requirements.txt
absl-py==0.6.1
adblockparser==0.7
alabaster==0.7.12
altgraph==0.17
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.2
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.7.1
astroid==2.1.0
astropy==3.2.1
atomicwrites==1.2.1
attrs==18.2.0
Automat==0.7.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.3
bitarray==0.8.3
bkcharts==0.2
blaze==0.11.3
bleach==3.0.2
bokeh==1.3.4
boto==2.49.0
boto3==1.9.253
botocore==1.12.253
bottle==0.12.18
Bottleneck==1.2.1
branca==0.3.1
business-calendar==0.2.1
certifi==2019.9.11
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
clyent==1.2.2
colorama==0.4.0
comtypes==1.1.7
conda==4.7.12
conda-build==3.17.1
conda-package-handling==1.6.0
configparser==4.0.2
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.7
cssselect==1.1.0
cycler==0.10.0
Cython==0.29
cytoolz==0.9.0.1
dask==1.0.0
datashape==0.5.4
decorator==4.3.0
defusedxml==0.5.0
dill==0.3.1.1
distributed==1.25.0
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.10
findspark==1.3.0
flask==1.0.2
flask_cors==3.0.7
folium==0.10.1
funcparserlib==0.3.6
future==0.18.2
gast==0.2.0
gensim==3.8.1
gevent==1.3.7
glob2==0.6
greenlet==0.4.15
grpcio==1.16.1
gunicorn==20.0.4
h5py==2.8.0
heapdict==1.0.0
html5lib==1.0.1
hyperlink==18.0.0
idna==2.7
imageio==2.4.1
imagesize==1.1.0
imbalanced-learn==0.5.0
imblearn==0.0
importlib-metadata==0.6
incremental==17.5.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.1
jieba==0.39
Jinja2==2.10
jmespath==0.9.4
joblib==0.14.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.3
jupyterlab-launcher==0.13.1
jupyterlab-server==0.2.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==16.1.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
libarchive-c==2.8
llvmlite==0.26.0
locket==0.2.0
lxml==4.2.5
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
menuinst==1.4.14
mido==1.2.9
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.1
modin==0.6.2
more-itertools==4.3.0
mpmath==1.0.0
msgpack==0.5.6
multipledispatch==0.6.0
mysql-connector-python==8.0.13
navigator-updater==0.2.1
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
nltk==3.3
nose==1.3.7
notebook==5.7.2
numba==0.41.0
numexpr==2.6.8
numpy==1.15.4
numpydoc==0.8.0
odo==0.5.1
olefile==0.46
openpyxl==2.5.11
packaging==18.0
pandas==0.25.1
pandas-datareader==0.8.1
pandocfilters==1.4.2
parsel==1.5.2
parso==0.3.1
partd==0.3.9
path.py==11.5.0
pathlib2==2.3.2
patsy==0.5.1
pefile==2019.4.18
pep8==1.7.1
pickleshare==0.7.5
Pillow==5.3.0
pkginfo==1.4.2
pluggy==0.8.0
ply==3.11
prometheus-client==0.4.2
prompt-toolkit==2.0.7
protobuf==3.6.1
psutil==5.4.8
py==1.7.0
py4j==0.10.7
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.3
PyDispatcher==2.0.5
pyflakes==2.0.0
Pygments==2.2.0
PyHamcrest==1.9.0
PyInstaller==3.6
pylint==2.2.2
pymongo==3.10.1
PyMySQL==0.9.3
pyod==0.7.5
pyodbc==4.0.24
pyOpenSSL==18.0.0
pyparsing==2.3.0
PySocks==1.6.8
pyspark==2.4.4
pytest==4.0.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.3.0
pytest-openfiles==0.3.1
pytest-remotedata==0.3.1
python-dateutil==2.7.5
pytz==2018.7
PyWavelets==1.0.1
pywin32==223
pywin32-ctypes==0.2.0
pywinpty==0.5.4
PyYAML==3.13
pyzmq==17.1.2
qt5reactor==0.5
QtAwesome==0.5.3
qtconsole==4.4.3
QtPy==1.5.2
queuelib==1.5.0
requests==2.20.1
rope==0.11.0
ruamel-yaml==0.15.46
s3transfer==0.2.1
scikit-image==0.14.0
scikit-learn==0.21.3
scipy==1.1.0
Scrapy==1.7.3
scrapy-splash==0.7.2
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
smart-open==1.8.4
snowballstemmer==1.2.1
sortedcollections==1.0.1
sortedcontainers==2.1.0
Sphinx==1.8.2
sphinxcontrib-websupport==1.1.0
splash==3.3.1
spyder==3.3.2
spyder-kernels==0.3.0
SQLAlchemy==1.2.14
statsmodels==0.10.1
sympy==1.3
tables==3.4.4
tblib==1.3.2
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
textrank==0.1.0
textrank4zh==0.3
toolz==0.9.0
tornado==5.1.1
tqdm==4.38.0
traitlets==4.3.2
Twisted==18.9.0
typed-ast==1.1.0
unicodecsv==0.14.1
urllib3==1.23
utils==0.9.0
w3lib==1.21.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
xgboost==0.90
xlrd==1.1.0
XlsxWriter==1.1.2
xlwings==0.15.0
xlwt==1.3.0
xvfbwrapper==0.2.9
zict==0.1.3
zope.interface==4.6.0
